backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.1.1
redis==5.0.1
celery==5.3.4
requests==2.32.4
pydantic==2.11.7
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
websockets==15.0.1
aiofiles==23.2.1
python-dateutil==2.8.2
pandas==2.0.3
numpy==1.24.3
pandas-ta==0.3.14b0
ccxt==4.1.77
pytz
apscheduler==3.10.4
schedule==1.2.0
